add plugin submission
Squashed commit of the following:
commit 2cb80542a01fc3aa5852d8c506dff9c36c7c7572
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 20:20:55 2021 +0200

    change status to pending in plugin dataTable

commit 1013c3f9bd55bcb6d46712ecc3e8701f654c52f4
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 20:16:19 2021 +0200

    disabble plugin form inputs when isPending

commit 342fb7f3f558cbea396d275e743cafe6545560cd
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 19:54:54 2021 +0200

    block requests on pending plugin

commit 8621b1f04b4631c117244912d04d326669975e6f
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 19:47:52 2021 +0200

    add the plugin submission form

commit 4236c409811a385448ddc43a35c685fbecd14885
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 19:24:29 2021 +0200

    disable file form when request is loading

commit 9f8e80a9aab02e3720c0220e849abb4009852ea4
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 18:55:08 2021 +0200

    add isLoading to useRequest

commit 2b535e43d1fefb66716930dcdf47d238832b6adc
Author: Karol Sitarz <ka.sitarz@gmail.com>
Date:   Wed Oct 27 18:43:23 2021 +0200

    change pluginStatus to isPending

+++ pages/api/plugins/[[...params]].ts
@@ -95,6 +95,7 @@ class Plugins {
       where: { id },
       include: {
         source: true,
+        published: true,
       },
     })
 
@@ -134,7 +135,11 @@ class Plugins {
     @Body(ValidationPipe) body: UpdatePluginDTO,
     @User user: RequestUser,
   ) {
-    await this.innerGetPlugin(id, user)
+    const plugin = await this.innerGetPlugin(id, user)
+
+    if (plugin.isPending) {
+      throw new BadRequestException(`Cannot make changes to a pending plugin.`)
+    }
 
     if (body?.minWidth && body?.maxWidth && body.minWidth > body.maxWidth) {
       throw new BadRequestException(
@@ -156,10 +161,59 @@ class Plugins {
 
   @Delete('/:id')
   async deletePlugin(@Param('id') id: string, @User user: RequestUser) {
-    await this.innerGetPlugin(id, user)
+    const plugin = await this.innerGetPlugin(id, user)
+
+    if (plugin.isPending) {
+      throw new BadRequestException(`Cannot make changes to a pending plugin.`)
+    }
 
     return await prisma.draftPlugin.delete({
       where: { id },
+    })
+  }
+
+  @Post('/:id/submit')
+  async submitPlugin(@Param('id') id: string, @User user: RequestUser) {
+    const plugin = await prisma.draftPlugin.findFirst({
+      where: { id },
+      include: {
+        source: true,
+      },
+    })
+
+    if (!plugin) {
+      throw new NotFoundException('The plugin does not exist.')
+    }
+
+    if (plugin.isPending) {
+      throw new BadRequestException(`Cannot make changes to a pending plugin.`)
+    }
+
+    if (plugin.authorId !== user.id) {
+      throw new UnauthorizedException('The plugin has a different author.')
+    }
+
+    if (!plugin.source) {
+      throw new BadRequestException(`Plugin must have the source code.`)
+    }
+
+    if (plugin.minWidth > plugin.maxWidth) {
+      throw new BadRequestException(
+        `Min. width can't be larger than max. width.`,
+      )
+    }
+
+    if (plugin.minHeight > plugin.maxHeight) {
+      throw new BadRequestException(
+        `Min. height can't be larger than max. height.`,
+      )
+    }
+
+    return await prisma.draftPlugin.update({
+      where: { id },
+      data: {
+        isPending: true,
+      },
     })
   }
 }
